fix: organize add-ons, add link color
