+++ packages/components/.storybook/addons/bug-hint/constants.ts
@@ -1,0 +1,3 @@
+export const ADDON_ID_BUG_HINT = 'solid/bug-hint';
+export const TOOL_ID = `${ADDON_ID_BUG_HINT}/tool`;
+export const PARAM_KEY = `solidBugHint`;
