+++ packages/components/.storybook/addons/theme-generator/preview.ts
@@ -1,5 +1,5 @@
 import type { Renderer, ProjectAnnotations } from '@storybook/types';
-import { PARAM_KEY } from '../constants';
+import { PARAM_KEY } from './constants';
 import { withGlobals } from './withGlobals';
 
 const preview: ProjectAnnotations<Renderer> = {
